move PacketBuilder class out of Paket to it's own file
the Packet and PacketBuilder classes are both big so moving them to different files are better
