+++ src/test/java/org/example/network/ChannelTest.java
@@ -1,9 +1,6 @@
 package org.example.network;
 
-import org.example.data.Flag;
-import org.example.data.Message;
-import org.example.data.Packet;
-import org.example.data.Payload;
+import org.example.data.*;
 import org.example.network.Channel;
 import org.example.network.interfaces.Endpoint;
 import org.example.network.interfaces.NetworkNode;
@@ -58,7 +55,7 @@ public class ChannelTest {
     public void channelPacketDeliversPacketToDestinationNodeTest(){
         Channel channel = new Channel(source, destination, rand, 100);
         Payload payload = new Message( "Test");
-        Packet packet = new Packet.PacketBuilder()
+        Packet packet = new PacketBuilder()
                 .withOrigin(source)
                 .withDestination(destination)
                 .withPayload(payload)
@@ -74,7 +71,7 @@ public class ChannelTest {
     public void lossyChannelDropPacketTest(){
         Channel channel = new Channel(source, destination, rand, 0);
         Payload payload = new Message( "Test");
-        Packet packet = new Packet.PacketBuilder()
+        Packet packet = new PacketBuilder()
                 .withOrigin(source)
                 .withDestination(destination)
                 .withPayload(payload)
@@ -91,7 +88,7 @@ public class ChannelTest {
         for (int i = 0; i < 10000; i++) {
             Channel channel = new Channel(source, destination, rand, 3);
             Payload payload = new Message( "Test");
-            Packet packet = new Packet.PacketBuilder()
+            Packet packet = new PacketBuilder()
                     .withOrigin(source)
                     .withDestination(destination)
                     .withPayload(payload)
