+++ src/test/java/org/example/network/RoutableTest.java
@@ -1,11 +1,8 @@
 package org.example.network;
 
-import org.example.data.Flag;
-import org.example.data.Message;
-import org.example.data.Payload;
+import org.example.data.*;
 import org.example.network.interfaces.Endpoint;
 import org.example.network.Router;
-import org.example.data.Packet;
 
 import org.example.protocol.BasicTCP;
 import org.junit.Assert;
@@ -63,7 +60,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -109,7 +106,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -151,7 +148,7 @@ public class RoutableTest {
         r3.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -194,7 +191,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -241,7 +238,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -285,7 +282,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -365,7 +362,7 @@ public class RoutableTest {
         r12.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -414,7 +411,7 @@ public class RoutableTest {
         r3.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -450,7 +447,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -487,7 +484,7 @@ public class RoutableTest {
         r4.start();
 
         Message msg = new Message( "hello på do!");
-        client.route(new Packet.PacketBuilder()
+        client.route(new PacketBuilder()
                 .withPayload(msg)
                 .withDestination(server)
                 .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
@@ -528,7 +525,7 @@ public class RoutableTest {
 
         for (int i = 0; i < 1000; i++) {
             Message msg = new Message( "hello på do!");
-            client.route(new Packet.PacketBuilder()
+            client.route(new PacketBuilder()
                     .withPayload(msg)
                     .withDestination(server)
                     .withFlags(Flag.SYN) // hack to overcome connection check in the endpoints
